backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
kafka-python==2.0.2
cryptography==41.0.7
python-dotenv==1.0.0
python-multipart==0.0.6

